==> 00-preface.Rmd <==
# (PART) DATA PREPARATION {-}

# Preparing for Downstream Analysis {#preface}

In this preface, we set the stage for an in-depth exploration into the complexities of preparing microbiome data. As the foundation of downstream analyses, effective data preparation is essential for extracting meaningful biological insights. This guide not only addresses the technical aspects of data cleaning, transformation, and integration but also offers practical insights to guide researchers through the nuanced decisions involved in optimizing their microbiome datasets. Whether you're a novice or an experienced researcher, embark on this journey to elevate the quality and reliability of your microbiome analysis.
